Add update stack segmentation as eventListener
