+++ packages/tools/src/eventListeners/segmentation/index.ts
@@ -2,10 +2,12 @@ import segmentationRepresentationModifiedEventListener from './segmentationRepre
 import segmentationDataModifiedEventListener from './segmentationDataModifiedEventListener';
 import segmentationRepresentationRemovedEventListener from './segmentationRepresentationRemovedEventListener';
 import segmentationModifiedListener from './segmentationModifiedEventListener';
+import stackImageChangeEventListener from './stackImageChangeEventListener';
 
 export {
   segmentationRepresentationModifiedEventListener,
   segmentationDataModifiedEventListener,
   segmentationRepresentationRemovedEventListener,
   segmentationModifiedListener,
+  stackImageChangeEventListener,
 };
